Printint mechanism improvement. - Make runner/printer (with result) identify actions by path on which they appear not by objects themselves. (WIP) Refactoring: - Split Action class into pieces.
