+++ src/test/java/com/github/dakusui/actionunit/actions/TimeoutTest.java
@@ -1,6 +1,5 @@
 package com.github.dakusui.actionunit.actions;
 
-import com.github.dakusui.actionunit.Action;
 import org.junit.Test;
 
 import static com.github.dakusui.actionunit.Actions.nop;
@@ -8,6 +7,6 @@ import static com.github.dakusui.actionunit.Actions.nop;
 public class TimeoutTest {
   @Test(expected = IllegalArgumentException.class)
   public void givenNegativeDuration$whenCreated$thenExceptionThrown() {
-    new Action.TimeOut(nop(), -2);
+    new TimeOut(nop(), -2);
   }
 }
